requirements.txt
Django==5.2.1
djangorestframework==3.16.0     
drf-yasg==1.21.10